+++ src/main/java/com/floriculturamonteiro/floricultura/repositories/FloresRepository.java
@@ -3,6 +3,6 @@ package com.floriculturamonteiro.floricultura.repositories;
 import com.floriculturamonteiro.floricultura.model.Flores;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FloresRepository extends JpaRepository<Flores, Integer> {
+public interface FloresRepository extends JpaRepository<Flores, Long> {
 
 }
